NU-2405 provide handleGenerateTestData
